+++ src/components/page/PagesConfig.ts
@@ -1,9 +1,17 @@
 ﻿module shiva {
     export interface PagesConfig {
-        pages: {},
-        id?: string,
-        style?: StyleDeclaration,
-        delayTransition?: number,
-        routes?: boolean
+        pages: {};
+        id?: string;
+        style?: StyleDeclaration;
+        delayTransition?: number;
+        routes?: boolean;
+        /**
+         * Specify an error view if address not found 
+        */
+        errorPage?: string;
+        /**
+         * Redirect to / if address not found
+        */
+        redirect?: boolean;
     }
 }